Move Prost and protobuf mods into their own files
